research-template/requirements.txt
natsort==8.1.0
scipy==1.8.0
pandas==1.4.0
matplotlib==3.5.1
torchmetrics==0.7.2
sklearn